+++ onchain/src/TrustlessSidechain/DistributedSet.hs
@@ -3,6 +3,7 @@
 {-# LANGUAGE MonoLocalBinds #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -fexpose-all-unfoldings #-}
 
 {- | Implementation of a set for on-chain proof of not in a set membership.
  We call this a *distributed set* since the set structure is distributed over
@@ -33,6 +34,7 @@ module TrustlessSidechain.DistributedSet (
   insertNode,
 
   -- * Validators / minting policies
+  mkInsertValidator,
   mkDsConfValidator,
   mkDsConfPolicy,
   dsConfTokenName,
@@ -410,6 +412,8 @@ mkDsConfPolicy dsc _red ctx =
 -}
 dsConfTokenName :: TokenName
 dsConfTokenName = TokenName emptyByteString
+
+{-# INLINE mkDsKeyPolicy #-}
 
 -- | 'mkDsKeyPolicy'.  See Note [How This All Works].
 mkDsKeyPolicy :: DsKeyMint -> () -> ScriptContext -> Bool