+++ onchain/src/TrustlessSidechain/UpdateCommitteeHash.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
+{-# OPTIONS_GHC -fexpose-all-unfoldings #-}
 
 module TrustlessSidechain.UpdateCommitteeHash where